Move object to new home
