+++ source/Energinet.DataHub.MeteringPoints.Application/Energinet.DataHub.MeteringPoints.Application.csproj
@@ -35,4 +35,8 @@ limitations under the License.
     <Compile Remove="MasterDataDocument.cs" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="GridAreas" />
+  </ItemGroup>
+
 </Project>
