+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/DataAccess/GridAreas/GridAreaRepository.cs
@@ -14,7 +14,6 @@
 
 using System;
 using System.Threading.Tasks;
-using Energinet.DataHub.MeteringPoints.Application.GridAreas;
 using Energinet.DataHub.MeteringPoints.Domain.GridAreas;
 using Microsoft.EntityFrameworkCore;
